APIs/stream-io-api.com/v80.2.0/python-aiohttp-client/openapi_server/controllers/imports_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server.models.api_error import APIError
from openapi_server.models.create_import_request import CreateImportRequest
from openapi_server.models.create_import_response import CreateImportResponse
from openapi_server.models.create_import_url_request import CreateImportURLRequest
from openapi_server.models.create_import_url_response import CreateImportURLResponse
from openapi_server.models.get_import_response import GetImportResponse
from openapi_server.models.list_imports_response import ListImportsResponse
from openapi_server import util


async def create_import(request: web.Request, body) -> web.Response:
    """Create import

    Creates a new import 

    :param body: 
    :type body: dict | bytes

    """
    body = CreateImportRequest.from_dict(body)
    return web.Response(status=200)


async def create_import_url(request: web.Request, body) -> web.Response:
    """Create import URL

    Creates a new import URL 

    :param body: 
    :type body: dict | bytes

    """
    body = CreateImportURLRequest.from_dict(body)
    return web.Response(status=200)


async def get_import(request: web.Request, id) -> web.Response:
    """Get import

    Gets an import 

    :param id: 
    :type id: str

    """
    return web.Response(status=200)


async def list_imports(request: web.Request, ) -> web.Response:
    """Get import

    Gets an import 


    """
    return web.Response(status=200)
