APIs/stream-io-api.com/v80.2.0/python-aiohttp-client/tests/test_imports_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.api_error import APIError
from openapi_server.models.create_import_request import CreateImportRequest
from openapi_server.models.create_import_response import CreateImportResponse
from openapi_server.models.create_import_url_request import CreateImportURLRequest
from openapi_server.models.create_import_url_response import CreateImportURLResponse
from openapi_server.models.get_import_response import GetImportResponse
from openapi_server.models.list_imports_response import ListImportsResponse


pytestmark = pytest.mark.asyncio

async def test_create_import(client):
    """Test case for create_import

    Create import
    """
    body = {"mode":"insert","path":"path"}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'stream-auth-type': 'special-key',
        'api_key': 'special-key',
        'JWT': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/imports',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_create_import_url(client):
    """Test case for create_import_url

    Create import URL
    """
    body = {"filename":"filename"}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'stream-auth-type': 'special-key',
        'api_key': 'special-key',
        'JWT': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/import_urls',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_import(client):
    """Test case for get_import

    Get import
    """
    headers = { 
        'Accept': 'application/json',
        'stream-auth-type': 'special-key',
        'api_key': 'special-key',
        'JWT': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/imports/{id}'.format(id='id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_imports(client):
    """Test case for list_imports

    Get import
    """
    headers = { 
        'Accept': 'application/json',
        'stream-auth-type': 'special-key',
        'api_key': 'special-key',
        'JWT': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/imports',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

